DJ_A2_02: Updated routes.ts
