--- src/router.tsx
@@ -1,37 +1,0 @@
-import { lazy } from 'react';
-
-import { createBrowserRouter, Navigate, RouteObject } from 'react-router-dom';
-
-import { MainLayout } from '@layout';
-
-const Dashboard = lazy(() =>
-    import('./pages').then((module) => ({
-        default: module.Dashboard,
-    })),
-);
-
-const NotFoundPage = lazy(() =>
-    import('./pages').then((module) => ({
-        default: module.NotFound,
-    })),
-);
-const ErrorPage = lazy(() =>
-    import('./pages').then((module) => ({
-        default: module.ErrorPage,
-    })),
-);
-
-const routes: RouteObject[] = [
-    {
-        path: '/',
-        errorElement: <Navigate to="/error" />,
-        element: <MainLayout />,
-        children: [
-            { index: true, element: <Dashboard /> },
-            { path: '/error', element: <ErrorPage /> },
-            { path: '*', element: <NotFoundPage /> },
-        ],
-    },
-];
-
-export const router = createBrowserRouter(routes);
